Quick hints to build a CentOS 7 image.
It was frustrating to try to start quickly via the Quick Start and not
have any idea where to find an image.  There are hints in the README in
the repo, but it's convenient to have them here.

I pulled the content from the README.md and referenced it and the
tutorials section.

"Someone" should probably work up a posse of examples for the Linux
Flavors du jour.